requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
importlib-metadata==4.0.1
Markdown==3.3.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.4.1
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
factory-boy==3.2.0